backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
yfinance
nsepython
apscheduler
email-validator
email-validator

